중간 저장

+++ PathHack/Models/Badge.swift
@@ -9,6 +9,6 @@ import Foundation
 
 struct Badge: Hashable {
     let name: String
-    let percent: Int
+    let percent: Double
     let imageName: String
 }
